Move single-used code to corresponding entity
